Add Pro version of tool for TomTom; bug fixes; repo reorg
Add the ArcGIS Pro version of the tool to process TomTom MultiNet data.

Reorganize the repo to separate the 10.x version and the Pro version.

Add a parameter to the 10.x version of the tool to choose whether to build the network dataset after creating it. https://github.com/Esri/street-data-processing-tools/issues/23

Fix https://github.com/Esri/street-data-processing-tools/issues/43